Refactoring hungryman package to wherefordinner in search app.
