+++ where-for-dinner/where-for-dinner-search/src/main/java/com/java/example/tanzu/wherefordinner/functions/SearchSupplier.java
@@ -1,4 +1,4 @@
-package com.java.example.tanzu.hungryman.functions;
+package com.java.example.tanzu.wherefordinner.functions;
 
 import java.util.function.Supplier;
 
@@ -6,12 +6,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.function.context.PollableBean;
 import org.springframework.context.annotation.Configuration;
 
-import com.java.example.tanzu.hungryman.entity.Search;
-import com.java.example.tanzu.hungryman.repository.SearchRepository;
+import com.java.example.tanzu.wherefordinner.entity.Search;
+import com.java.example.tanzu.wherefordinner.repository.SearchRepository;
 
+import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
 
 @Configuration
+@Slf4j
 public class SearchSupplier
 {
 	@Autowired
@@ -22,7 +24,10 @@ public class SearchSupplier
 	{
 		return () ->
 		{
+			
 			final var curTime = System.currentTimeMillis();
+			
+			log.info("Gathering and sending all active searches to downstream processing");
 			
 			/*
 			 * If the requested end time has passed, then don't search
