+++ where-for-dinner/where-for-dinner-search/src/main/resources/application.yml
@@ -11,6 +11,7 @@ where-for-dinner:
     
 spring: 
   sql.init.mode: ALWAYS 
+  sql.init.platform: h2
 
   cloud: 
     stream:
@@ -23,7 +24,7 @@ spring:
           destination: where-for-dinner-search-criteria
 
 springdoc:
-  packages-to-scan: com.java.example.tanzu.hungryman.resources
+  packages-to-scan: com.java.example.tanzu.wherefordinner.resources
   show-spring-cloud-functions: false
           
 ---
@@ -31,7 +32,16 @@ springdoc:
 spring:
   config.activate.on-profile: mysql
   
+  sql.init.platform: mysql
+  
   r2dbc: 
     #Default properties to work with MariaDB driver
     properties.sslMode: TRUST  
     properties.tlsVersion: TLSv1.2
+    
+---
+
+spring:
+  config.activate.on-profile: postgres
+  
+  sql.init.platform: postgresql
