+++ where-for-dinner/where-for-dinner-search/src/main/resources/schema-h2.sql
